fix: add missing beta NoOpFeatureFlagOverrides
