+++ app-k9mail/src/release/kotlin/app/k9mail/dev/ReleaseConfig.kt
@@ -14,5 +14,5 @@ fun Module.developmentModuleAdditions() {
     single<List<AutoDiscovery>>(named("extraAutoDiscoveries")) {
         emptyList()
     }
-    single<FeatureFlagOverrides> { NoOpFeatureFlagOverrides }
+    single<FeatureFlagOverrides> { NoOpFeatureFlagOverrides() }
 }
