+++ app-thunderbird/src/beta/kotlin/net/thunderbird/android/dev/BetaConfig.kt
@@ -2,6 +2,8 @@ package net.thunderbird.android.dev
 
 import app.k9mail.autodiscovery.api.AutoDiscovery
 import com.fsck.k9.backend.BackendFactory
+import net.thunderbird.core.featureflag.FeatureFlagOverrides
+import net.thunderbird.core.featureflag.NoOpFeatureFlagOverrides
 import org.koin.core.module.Module
 import org.koin.core.qualifier.named
 
@@ -12,4 +14,6 @@ fun Module.developmentModuleAdditions() {
     single<List<AutoDiscovery>>(named("extraAutoDiscoveries")) {
         emptyList()
     }
+
+    single<FeatureFlagOverrides> { NoOpFeatureFlagOverrides() }
 }
