+++ app-thunderbird/src/release/kotlin/net/thunderbird/android/dev/ReleaseConfig.kt
@@ -14,5 +14,5 @@ fun Module.developmentModuleAdditions() {
     single<List<AutoDiscovery>>(named("extraAutoDiscoveries")) {
         emptyList()
     }
-    single<FeatureFlagOverrides> { NoOpFeatureFlagOverrides }
+    single<FeatureFlagOverrides> { NoOpFeatureFlagOverrides() }
 }
